+++ src/dictionarly/components/dictionarly-end-modal/dictionarly-end-modal.styled.ts
@@ -1,9 +1,16 @@
-import * as R from 'ramda'
 import styled, { ThemeProps } from 'styled-components/native'
+import { Modalize } from 'react-native-modalize'
 import { getThemeProp } from '@core/styled/theme'
 import { TEXT_SIZE } from '@core/text/text.constants'
 import { COLOR } from '@core/colors/colors.constants'
 import { ThemeModel } from '@core/styled/models'
+
+export const DictionarlyEndModalize = styled(Modalize).attrs(props => ({
+  adjustToContentHeight: true,
+  modalStyle: {
+    backgroundColor: getThemeProp('backgroundPrimary')(props)
+  }
+}))``
 
 export const DictionarlyEndModalContainer = styled.View`
   height: 300px;